Tidy up
- Invert netsonly and filenamesonly options.
- Change start/end to from/to options.
- Add option to select boost GraphViz parser.
- Add asserts to new parser.
- Add a unit test for boost parser.
- Other formatting.

+++ include/netlist_paths/Vertex.hpp
@@ -29,7 +29,7 @@ enum VertexDirection {
 };
 
 struct VertexProperties {
-  int id;
+  unsigned long long id;
   VertexType type;
   VertexDirection dir;
   std::string name;
